+++ libs/mini-rx-store/src/index.ts
@@ -24,3 +24,4 @@ export { LoggerExtension } from './lib/extensions/logger.extension';
 export { ImmutableStateExtension } from './lib/extensions/immutable-state.extension';
 export { UndoExtension, undo } from './lib/extensions/undo.extension';
 export { tapResponse } from './lib/tap-response';
+export { mapResponse } from './lib/map-response';